followingq

+++ src/main/java/server/service/following/FollowingService.java
@@ -4,10 +4,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import server.domain.album.Album;
 import server.domain.following.FollowingInfo;
-import server.domain.melody_card.MelodyCard;
 import server.domain.member.persist.Member;
+import server.domain.member.vo.MemberSession;
 import server.global.exception.BadRequestException;
-import server.mapper.card_favorite.dto.CardFavoriteRequest;
 import server.mapper.following.dto.FollowingDto;
 import server.mapper.following.dto.FollowingInfoDto;
 import server.mapper.following.dto.FollowingResponse;
@@ -21,6 +20,8 @@ import server.service.notification.NotificationCreateService;
 import java.util.ArrayList;
 import java.util.List;
 
+import static server.global.constant.ExceptionMessage.FOLLOWING_ALREADY_EXISTS;
+import static server.global.constant.ExceptionMessage.FOLLOWING_NOT_FOUND;
 import static server.global.constant.NotificationConstant.FOLLOWING_NOTIFICATION;
 
 @Service
@@ -44,9 +45,20 @@ public class FollowingService {
 
     @Transactional
     public void saveFollowing(Long memberId, FollowingDto followingDto){
+
+        FollowingInfo following = followingRepository.findByFollowingIdAndFollowedId(memberId,followingDto.memberId());
+
+        if(following != null){
+            throw new BadRequestException(FOLLOWING_ALREADY_EXISTS.message);
+        }
+
         FollowingInfo followingInfo = FollowingInfo.builder().followedId(followingDto.memberId()).followingId(memberId).build();
         followingRepository.save(followingInfo);
         notification(memberId, followingDto);
+
+
+
+
     }
 
     private void notification(final long memberId, final FollowingDto followingDto) {
@@ -58,6 +70,9 @@ public class FollowingService {
     @Transactional
     public void deleteFollowing(FollowingInfoDto followingInfoDto){
         FollowingInfo followingInfo = followingRepository.findByFollowingIdAndFollowedId(followingInfoDto.followingId(), followingInfoDto.followedId());
+        if(followingInfo == null){
+            throw new BadRequestException(FOLLOWING_NOT_FOUND.message);
+        }
         followingRepository.delete(followingInfo);
     }
 
@@ -119,5 +134,17 @@ public class FollowingService {
     }
 
 
+    @Transactional
+    public boolean getFollowing(MemberSession memberSession, Long memberId){
+        try {
+            FollowingInfo followingInfo = followingRepository.findByFollowingIdAndFollowedId(memberSession.id(),memberId);
+            if(followingInfo == null){
+                throw new BadRequestException(FOLLOWING_NOT_FOUND.message);
+            }
+            return true;
+        }catch (BadRequestException e){
+            return false;
+        }
+    }
 
 }
